set up notify config & client
