+++ src/config/server.js
@@ -1,0 +1,122 @@
+import path from 'node:path'
+import { fileURLToPath } from 'node:url'
+import environments from '../api/common/constants/environments.js'
+
+const dirname = path.dirname(fileURLToPath(import.meta.url))
+
+const isProduction = process.env.NODE_ENV === environments.PRODUCTION
+const isDev = process.env.NODE_ENV === environments.DEVELOPMENT
+const isTest = process.env.NODE_ENV === environments.TEST
+
+export const serverConfig = {
+  serviceVersion: {
+    doc: 'The service version, this variable is injected into your docker container in CDP environments',
+    format: String,
+    nullable: true,
+    default: null,
+    env: 'SERVICE_VERSION'
+  },
+  env: {
+    doc: 'The application environment.',
+    format: ['production', 'development', 'test'],
+    default: 'development',
+    env: 'NODE_ENV'
+  },
+  port: {
+    doc: 'The port to bind.',
+    format: 'port',
+    default: 3001,
+    env: 'PORT'
+  },
+  serviceName: {
+    doc: 'Api Service Name',
+    format: String,
+    default: 'fcp-sfd-comms'
+  },
+  root: {
+    doc: 'Project root',
+    format: String,
+    default: path.resolve(dirname, '../..')
+  },
+  isProduction: {
+    doc: 'If this application running in the production environment',
+    format: Boolean,
+    default: isProduction
+  },
+  isDevelopment: {
+    doc: 'If this application running in the development environment',
+    format: Boolean,
+    default: isDev
+  },
+  isTest: {
+    doc: 'If this application running in the test environment',
+    format: Boolean,
+    default: isTest
+  },
+  log: {
+    enabled: {
+      doc: 'Is logging enabled',
+      format: Boolean,
+      default: !isTest,
+      env: 'LOG_ENABLED'
+    },
+    level: {
+      doc: 'Logging level',
+      format: ['fatal', 'error', 'warn', 'info', 'debug', 'trace', 'silent'],
+      default: 'info',
+      env: 'LOG_LEVEL'
+    },
+    format: {
+      doc: 'Format to output logs in.',
+      format: ['ecs', 'pino-pretty'],
+      default: isProduction ? 'ecs' : 'pino-pretty',
+      env: 'LOG_FORMAT'
+    },
+    redact: {
+      doc: 'Log paths to redact',
+      format: Array,
+      default: isProduction
+        ? ['req.headers.authorization', 'req.headers.cookie', 'res.headers']
+        : ['req', 'res', 'responseTime']
+    }
+  },
+  mongoUri: {
+    doc: 'URI for mongodb',
+    format: String,
+    default: 'mongodb://127.0.0.1:27017/',
+    env: 'MONGO_URI'
+  },
+  mongoDatabase: {
+    doc: 'database for mongodb',
+    format: String,
+    default: 'fcp-sfd-comms',
+    env: 'MONGO_DATABASE'
+  },
+  httpProxy: {
+    doc: 'HTTP Proxy',
+    format: String,
+    nullable: true,
+    default: null,
+    env: 'HTTP_PROXY'
+  },
+  isSecureContextEnabled: {
+    doc: 'Enable Secure Context',
+    format: Boolean,
+    default: isProduction,
+    env: 'ENABLE_SECURE_CONTEXT'
+  },
+  isMetricsEnabled: {
+    doc: 'Enable metrics reporting',
+    format: Boolean,
+    default: isProduction,
+    env: 'ENABLE_METRICS'
+  },
+  tracing: {
+    header: {
+      doc: 'Which header to track',
+      format: String,
+      default: 'x-cdp-request-id',
+      env: 'TRACING_HEADER'
+    }
+  }
+}
